Add editable role permission matrix

+++ routes/web.php
@@ -20,20 +20,18 @@ Route::middleware('guest')->group(function (): void {
 
 Route::middleware('auth')->group(function (): void {
     Route::post('/logout', [AuthController::class, 'destroy'])->name('logout');
-    Route::get('/', [ProjectDashboardController::class, 'index'])->name('dashboard');
-    Route::get('/tugas-saya', [MyTaskController::class, 'index'])->name('my-tasks.index');
+    Route::get('/', [ProjectDashboardController::class, 'index'])->middleware('permission:dashboard_view')->name('dashboard');
+    Route::get('/tugas-saya', [MyTaskController::class, 'index'])->middleware('permission:process_view')->name('my-tasks.index');
 
     Route::prefix('projects')->name('projects.')->group(function (): void {
-        Route::middleware('role:admin,manager')->group(function (): void {
-            Route::get('/create', [ProjectController::class, 'create'])->name('create');
-            Route::post('/', [ProjectController::class, 'store'])->name('store');
-            Route::get('/{project}/edit', [ProjectController::class, 'edit'])->name('edit');
-            Route::put('/{project}', [ProjectController::class, 'update'])->name('update');
-            Route::delete('/{project}', [ProjectController::class, 'destroy'])->name('destroy');
-        });
+        Route::get('/create', [ProjectController::class, 'create'])->middleware('permission:project_create')->name('create');
+        Route::post('/', [ProjectController::class, 'store'])->middleware('permission:project_create')->name('store');
+        Route::get('/{project}/edit', [ProjectController::class, 'edit'])->middleware('permission:project_update')->name('edit');
+        Route::put('/{project}', [ProjectController::class, 'update'])->middleware('permission:project_update')->name('update');
+        Route::delete('/{project}', [ProjectController::class, 'destroy'])->middleware('permission:project_delete')->name('destroy');
 
-        Route::get('/{project}', [ProjectDashboardController::class, 'show'])->name('show');
-        Route::get('/{project}/processes/{process}', [ProjectDashboardController::class, 'showProcess'])->name('processes.show');
+        Route::get('/{project}', [ProjectDashboardController::class, 'show'])->middleware('permission:project_view')->name('show');
+        Route::get('/{project}/processes/{process}', [ProjectDashboardController::class, 'showProcess'])->middleware('permission:process_view')->name('processes.show');
         Route::post('/{project}/processes/{process}/checklists', [ProjectProcessChecklistController::class, 'store'])->name('processes.checklists.store');
         Route::put('/{project}/processes/{process}/checklists/{checklist}', [ProjectProcessChecklistController::class, 'update'])->name('processes.checklists.update');
         Route::delete('/{project}/processes/{process}/checklists/{checklist}', [ProjectProcessChecklistController::class, 'destroy'])->name('processes.checklists.destroy');
@@ -41,7 +39,7 @@ Route::middleware('auth')->group(function (): void {
         Route::delete('/{project}/processes/{process}/comments/{comment}', [ProjectProcessCommentController::class, 'destroy'])->name('processes.comments.destroy');
     });
 
-    Route::prefix('master-flows')->name('master-flows.')->middleware('role:admin')->group(function (): void {
+    Route::prefix('master-flows')->name('master-flows.')->middleware('permission:master_flow_manage')->group(function (): void {
         Route::get('/', [MasterFlowController::class, 'index'])->name('index');
         Route::post('/', [MasterFlowController::class, 'store'])->name('store');
         Route::get('/{masterFlow}/edit', [MasterFlowController::class, 'edit'])->name('edit');
@@ -58,14 +56,15 @@ Route::middleware('auth')->group(function (): void {
         Route::delete('/{masterFlow}/steps/{step}/checklists/{checklist}', [MasterFlowController::class, 'destroyChecklist'])->name('steps.checklists.destroy');
     });
 
-    Route::prefix('roles')->name('roles.')->middleware('role:admin')->group(function (): void {
+    Route::prefix('roles')->name('roles.')->middleware('permission:role_manage')->group(function (): void {
         Route::get('/', [RoleController::class, 'index'])->name('index');
         Route::post('/', [RoleController::class, 'store'])->name('store');
+        Route::put('/permissions', [RoleController::class, 'updatePermissions'])->name('permissions.update');
         Route::put('/{role}', [RoleController::class, 'update'])->name('update');
         Route::delete('/{role}', [RoleController::class, 'destroy'])->name('destroy');
     });
 
-    Route::prefix('users')->name('users.')->middleware('role:admin')->group(function (): void {
+    Route::prefix('users')->name('users.')->middleware('permission:user_manage')->group(function (): void {
         Route::get('/', [UserManagementController::class, 'index'])->name('index');
         Route::get('/create', [UserManagementController::class, 'create'])->name('create');
         Route::post('/', [UserManagementController::class, 'store'])->name('store');
